Revert bool/Bool implementation for the below mentioned Modules:
[KERNEL] 168a794f16611ed035572a5f762eed4682de5bbc.
[LPM] b0b358e538cb5f4ca789811104f30c860b34ba71.
[MCASP] 45ab3cde1d55d53fe485232b6da61e480166aa47.
[DSS] 12834e4c8356ba4e5def1ca0bd1162eaa21f253f.
[FATFS] 4ffe346cbaac27889735c06a394b4192de5fdb34.
[PCIE] 9f050385cca0190fd62072ed9f055a35e3283b77.
[MMCSD] 24d7fb2f0ed97548232bd5d9851862ad00fda3a1.
[OSPI] 48d37a010a7d9bd10a58666d96620c010092afa8.
[BOOT] 9ead181e998c40022c278e3d35bf8fbf4fa5cca6.
[UART] a8f8d2c5bf89a43d68ea6cdd303afeee4c558f3d.
[GPIO] 0cbf515ffe5cd03a773246aad9c2f1e2df6b1dad.
[SCICLIENT] f03d71041062bc1eface645578dc0c78549da64d.
[SCICLIENT] d16aa73f2c5ca5d5cccc04af156a4df9a722d254.
[UDMA] 52a05193ea58876972841210cafdf20efb0808f3.
[UDMA_UTILS] 33c50581ba869ef93e1c6346c97c82cd2888f34d.
[BOARD] b080da2ff6810fdff885b3cdcaac00c6c2ebcc3b.
[BOARD] 5fb6d7b418b17c47714e1fa0311af2b27b7641e9.
[BOARD] 691f99e27ed48234092937baeaee8e780b10e2d0.
[OSAL] fbe051cd3323133274a989018655ddb0b1f3b06c.
[I2C] 6e8661d9de64f2cf00095b2219ba9a7d2780f136.
[IPC] b0e2470359a073ff2f4aa98e34f453000a9fde2c.
[FVID2] 075b6d9268fa3790aa1b73779bc4349e99c40dd0.

+++ packages/ti/board/src/flash/nor/nor.h
@@ -62,16 +62,16 @@ typedef uintptr_t NOR_HANDLE;
 
 /** NOR return type macros */
 /** \brief On Success. */
-#define NOR_PASS                         (int32_t)(0)
+#define NOR_PASS                         (0U)
 
 /** \brief On Failure. */
-#define NOR_FAIL                         (int32_t)(-1U)
+#define NOR_FAIL                         (-1U)
 
 /** \brief Error code for Timeout. */
-#define NOR_TIMEOUT                      (int32_t)(-2U)
+#define NOR_TIMEOUT                      (-2U)
 
 /** \brief Error code to indicate invalid parameter. */
-#define NOR_INVALID_PARAM                (int32_t)(-3U)
+#define NOR_INVALID_PARAM                (-3U)
 
 /*
 * \breif NOR device size (bus width) in bits
@@ -280,8 +280,8 @@ NOR_STATUS NOR_write(NOR_HANDLE handle, uint32_t addr,
  * \param     handle  [IN]      : flash device handle
  * \param     sector  [IN]      : Sector number to be erased
  *                                if -1, device erase
- * \param     blkErase[IN]      : BFALSE - Sector erase
- *                                BTRUE  - Bloack erase
+ * \param     blkErase[IN]      : false - Sector erase
+ *                                true - Bloack erase
  *
  * \return    NOR_PASS      : if success.
  *            NOR_FAIL      : if error.
