+++ packages/ti/board/src/j7200_evm/include/board_utils.h
@@ -69,10 +69,10 @@ extern "C" {
 
 #define BOARD_MAC_ADDR_BYTES                  (6U)
 
-#define BOARD_ENET_NONE                       ((int32_t)(0))
-#define BOARD_ENET_QSGMII                     ((int32_t)(1))
-#define BOARD_ENET_SGMII                      ((int32_t)(2))
-#define BOARD_ENET_UNKOWN                     ((int32_t)(-1))
+#define BOARD_ENET_NONE                       (0)
+#define BOARD_ENET_QSGMII                     (1U)
+#define BOARD_ENET_SGMII                      (2U)
+#define BOARD_ENET_UNKOWN                     (-1)
 
 /**
  * \brief Structure to configure the board I2C parameters
@@ -142,10 +142,10 @@ typedef struct Board_DetectCfg_s
  * \n                      BOARD_ID_SOM(0x2) - Dual PMIC SoM Board
  * \n                      BOARD_ID_CP(0x3) - CP Board
  *
- * \return   BTRUE if the given board is detected else BFALSE.
+ * \return   TRUE if the given board is detected else FALSE.
  *           SoM board will be always connected to the base board.
- *           For SoM boardID return value BTRUE indicates dual PMIC
- *           SoM and BFALSE indicates alternate PMIC SoM
+ *           For SoM boardID return value TRUE indicates dual PMIC
+ *           SoM and FALSE indicates alternate PMIC SoM
  *
  */
 bool Board_detectBoard(uint32_t boardID);
@@ -160,7 +160,7 @@ bool Board_detectBoard(uint32_t boardID);
  * \n                      BOARD_ID_GESI(0x2) - GESI Board
  * \n                      BOARD_ID_ENET(0x3) - Quad ENET expansion
  *
- * \return BTRUE if board revision is E2, BFALSE for all other cases
+ * \return TRUE if board revision is E2, FALSE for all other cases
  */
 bool Board_isAlpha(uint32_t boardID);
