+++ packages/ti/board/src/j784s4_evm/include/board_utils.h
@@ -70,10 +70,10 @@ extern "C" {
 
 #define BOARD_MAC_ADDR_BYTES                  (6U)
 
-#define BOARD_ENET_NONE                       ((int32_t)(0))
-#define BOARD_ENET_QSGMII                     ((int32_t)(1))
-#define BOARD_ENET_SGMII                      ((int32_t)(2))
-#define BOARD_ENET_UNKOWN                     ((int32_t)(-1))
+#define BOARD_ENET_NONE                       (0)
+#define BOARD_ENET_QSGMII                     (1U)
+#define BOARD_ENET_SGMII                      (2U)
+#define BOARD_ENET_UNKOWN                     (-1)
 
 /* GPIO port and pin numbers for SDIO 1V8 enable */
 #define BOARD_SDIO_1V8_EN_PIN_NUM             (8U) //GPIO0_8 - SEL_SDIO_3V3_1V8n
@@ -169,10 +169,10 @@ Board_STATUS Board_getBoardData(Board_IDInfo_v2 *info, uint32_t boardID);
  * \n                      BOARD_ID_ENET(0x2) - ENET Board
  * \n                      BOARD_ID_EVM(0x3) - EVM Board
  *
- * \return   BTRUE if the given board is detected else BFALSE.
+ * \return   TRUE if the given board is detected else FALSE.
  *           SoM board will be always connected to the base board.
- *           For SoM boardID return value BTRUE indicates dual PMIC
- *           SoM and BFALSE indicates alternate PMIC SoM
+ *           For SoM boardID return value TRUE indicates dual PMIC
+ *           SoM and FALSE indicates alternate PMIC SoM
  *
  */
 bool Board_detectBoard(uint32_t boardID);
@@ -186,7 +186,7 @@ bool Board_detectBoard(uint32_t boardID);
  * \n                      BOARD_ID_ENET(0x2) - ENET Board
  * \n                      BOARD_ID_EVM(0x3) - EVM Board
  *
- * \return BTRUE if board revision is E2, BFALSE for all other cases
+ * \return TRUE if board revision is E2, FALSE for all other cases
  */
 bool Board_isAlpha(uint32_t boardID);
