+++ packages/ti/drv/sciclient/src/sciserver/j721e/sciserver_soc.c
@@ -48,7 +48,7 @@ static void J721E_ClearBootInterrupts (void);
 
 static void Sciserver_enableAcspcie (void)
 {
-    mmr_unlock((uint32_t)CSL_CTRL_MMR0_CFG0_BASE, 2U);
+    mmr_unlock(CSL_CTRL_MMR0_CFG0_BASE, 2);
 
     /**
      * Enable the control bit for CTRLMMR_PCIE_REFCLK0
@@ -74,7 +74,7 @@ static void Sciserver_enableAcspcie (void)
     addr=(uint32_t *)(CSL_CTRL_MMR0_CFG0_BASE + CSL_MAIN_CTRL_MMR_CFG0_PCIE_REFCLK3_CLKSEL);
     CSL_REG32_FINS(addr,MAIN_CTRL_MMR_CFG0_PCIE_REFCLK3_CLKSEL_OUT_CLK_EN,1);
 
-    mmr_lock((uint32_t)CSL_CTRL_MMR0_CFG0_BASE, 2U);
+    mmr_lock(CSL_CTRL_MMR0_CFG0_BASE, 2);
 }
 
 #define MAIN2MCU_LVL_INTRTR0_OUTL_32 (32U)
