+++ packages/ti/drv/sciclient/src/sciserver/j721s2/sciserver_soc.c
@@ -47,7 +47,7 @@ static void Sciserver_enableAcspcie (void);
 static void Sciserver_enableAcspcie (void)
 {
 
-    mmr_unlock((uint32_t)CSL_CTRL_MMR0_CFG0_BASE, 2U);
+    mmr_unlock(CSL_CTRL_MMR0_CFG0_BASE, 2);
 
     /**
      * Enable the control bit for CTRLMMR_PCIE_REFCLK1
@@ -55,7 +55,7 @@ static void Sciserver_enableAcspcie (void)
     uint32_t *addr=(uint32_t *)(CSL_CTRL_MMR0_CFG0_BASE + CSL_MAIN_CTRL_MMR_CFG0_PCIE_REFCLK1_CLKSEL);
     CSL_REG32_FINS(addr,MAIN_CTRL_MMR_CFG0_PCIE_REFCLK1_CLKSEL_OUT_CLK_EN,1);
 
-    mmr_lock((uint32_t)CSL_CTRL_MMR0_CFG0_BASE, 2U);
+    mmr_lock(CSL_CTRL_MMR0_CFG0_BASE, 2);
 }
 
 void Sciserver_socInit(void)
